LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-initial-target/scroll-initial-target-with-user-programmatic-scroll.tentative-expected.txt


FAIL scroll-initial-target does not override user scroll assert_approx_equals: scroll-initial-target sets initial vertical scroll position expected 200 +/- 1 but got 0
FAIL scroll-initial-target does not override programmatic scroll assert_approx_equals: scroll-initial-target sets initial vertical scroll position expected 200 +/- 1 but got 0
